% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyBar.R
\name{lazyProgressBar}
\alias{lazyProgressBar}
\title{Progress bar with customisable estimated remaining time}
\usage{
lazyProgressBar(n, method = "average", fn = NULL, ...)
}
\arguments{
\item{n}{Interger. Total number of ticks}

\item{method}{Character. The embeded forecasting method of remaining time:
\code{drift} (default), \code{average}, \code{naive}, or \code{snaive}.
Ignored if \code{fn} is not \code{NULL}.
\describe{
\item{\code{average} (default)}{Average method. The run time between future ticks are assumed to
be the average run time of the past ticks.
This is the most common estimation method for remaining time.}
\item{\code{drift}}{Drift method. The run time between future ticks are
assumned to increase (descrease), and the level changed is set to be the average change
of the run time of the past ticks.
This is to assume the run time between ticks is linearly increasing or decreasing.}
\item{\code{naive}}{Naive method. The run time between futuren ticks are assumed to be
the run time between the last two ticks,}
\item{\code{snaive}}{Seasonal naive method. If this method is chosen, an argument of \code{s}
needs to be supplied in the \code{...}.
The run time between future ticks is set to be the run time \code{s} times before.
By default \code{s} is set to be 1/10 of the total number of ticks.}
}}

\item{fn}{Function. User defined function to estimate the remaining time.
The function should predict the remaining time using the arguments and
return a scale.
It should have at least three arguments in the order of \code{dtime}, \code{i}, \code{n},
which represent the status of the pogressor bar at the current tick:
\describe{
\item{\code{dtime}} {A numeric vector of the run time between past ticks.}
\item{\code{i}} {The number of the current tick.}
\item{\code{n}} {The number of total ticks.}
}}

\item{...}{Other arguments to pass to estimation method.
The arguments need to be named.}
}
\value{
An R6 object with methods \code{tick()} and \code{print()}.
}
\description{
Display a progress bar displaying the estimated time.
The purpose of having various estimation mathods is to
provide a more accurate estimation when the run time between
ticks is assumed to be different, e.g., online estimation,
time series cross validation, expanding window approach, etc.
}
\details{
Four simple forecasting methods are available for
the estimation of the remaining time:
Average method (default), Drift method, Naive method and
Seasonal naive method.
User can also supply their customised estimation method as a function.
See \code{Arguments} and \code{Examples}.
}
